# doc/api.yml
openapi: 3.0.3
info:
  title: Cinema Reservation API
  version: 1.0.0
  description: API for reserving seats to cinema projections, with role-based access control.
security:
  - JwtAuth: []
paths:
  /movies:
    get:
      summary: List movies
      operationId: getMovies
      description: Accessible to everyone, no authentication required.
      security: []
      responses:
        '200':
          description: List of movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
    post:
      summary: Create a movie
      operationId: createMovie
      description: |
        Allowed for MANAGER only.
        - CLIENT, EMPLOYEE -> 403
        - MANAGER -> 201
      security:
        - JwtAuth: [MANAGER]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
      responses:
        '201':
          description: Movie created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
  /movies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a movie by id
      operationId: getMovieById
      description: Accessible to everyone, no authentication required.
      security: []
      responses:
        '200':
          description: The requested movie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a movie
      operationId: updateMovie
      description: |
        Allowed for MANAGER only.
        - CLIENT, EMPLOYEE -> 403
        - MANAGER -> 200
      security:
        - JwtAuth: [MANAGER]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
      responses:
        '200':
          description: Movie updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a movie
      operationId: deleteMovie
      description: |
        Allowed for MANAGER only.
        - CLIENT, EMPLOYEE -> 403
        - MANAGER -> 204
      security:
        - JwtAuth: [MANAGER]
      responses:
        '204':
          description: Movie deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projections:
    get:
      summary: List projections
      operationId: getProjections
      description: Accessible to everyone, no authentication required.
      security: []
      responses:
        '200':
          description: List of projections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Projection'
    post:
      summary: Create a projection
      operationId: createProjection
      description: |
        Allowed for MANAGER only.
        - CLIENT, EMPLOYEE -> 403
        - MANAGER -> 201
      security:
        - JwtAuth: [MANAGER]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Projection'
      responses:
        '201':
          description: Projection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projections/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a projection by id
      operationId: getProjectionById
      description: Accessible to everyone, no authentication required.
      security: []
      responses:
        '200':
          description: The requested projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projection'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a projection
      operationId: updateProjection
      description: |
        Allowed for MANAGER only.
        - CLIENT, EMPLOYEE -> 403
        - MANAGER -> 200
      security:
        - JwtAuth: [MANAGER]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Projection'
      responses:
        '200':
          description: Projection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a projection
      operationId: deleteProjection
      description: |
        Allowed for MANAGER only.
        - CLIENT, EMPLOYEE -> 403
        - MANAGER -> 204
      security:
        - JwtAuth: [MANAGER]
      responses:
        '204':
          description: Projection deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /reservations:
    get:
      summary: List reservations
      operationId: getReservations
      description: |
        Allowed for EMPLOYEE and MANAGER only.
        - CLIENT -> 403
        - EMPLOYEE, MANAGER -> 200
      security:
        - JwtAuth: [EMPLOYEE, MANAGER]
      responses:
        '200':
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Create a reservation
      operationId: createReservation
      description: |
        Allowed for EMPLOYEE and MANAGER only.
        - CLIENT -> 403
        - EMPLOYEE, MANAGER -> 201
      security:
        - JwtAuth: [EMPLOYEE, MANAGER]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
  /reservations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a reservation by id
      operationId: getReservationById
      description: |
        Authenticated users only.
        - CLIENT: 200 only if the reservation belongs to them, 403 otherwise
        - EMPLOYEE, MANAGER -> 200
      security:
        - JwtAuth: []
      responses:
        '200':
          description: The requested reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a reservation
      operationId: updateReservation
      description: |
        Allowed for EMPLOYEE and MANAGER only.
        - CLIENT -> 403
        - EMPLOYEE, MANAGER -> 200
      security:
        - JwtAuth: [EMPLOYEE, MANAGER]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '200':
          description: Reservation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a reservation
      operationId: deleteReservation
      description: |
        Allowed for EMPLOYEE and MANAGER only.
        - CLIENT -> 403
        - EMPLOYEE, MANAGER -> 204
      security:
        - JwtAuth: [EMPLOYEE, MANAGER]
      responses:
        '204':
          description: Reservation deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Movie:
      type: object
      required: [title, genres, duration]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        genres:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Genre'
        description:
          type: string
        duration:
          type: string
          description: ISO-8601 duration
    Projection:
      type: object
      required: [datetime, seatPrice, roomId, movieId]
      properties:
        id:
          type: string
          format: uuid
        datetime:
          type: string
          format: date-time
        seatPrice:
          type: number
          format: decimal
        roomId:
          type: string
          format: uuid
        movieId:
          type: string
          format: uuid
    Reservation:
      type: object
      required: [projectionId, userId, seatIds]
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        projectionId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        seatIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    Genre:
      type: string
      enum: [THRILLER, ROMANCE, COMEDY, DRAMA, ACTION, SCI_FI, FANTASY, ANIMATION]
  responses:
    BadRequest:
      description: Invalid request body
    Unauthorized:
      description: Missing or invalid authentication token
    Forbidden:
      description: Insufficient role or not the owner of the resource
    NotFound:
      description: Resource not found